refactor: move UserStarService into dedicated file
